requirements.txt
ansible==2.10.3
ansible-base==2.10.3
ansible-vault==1.2.0
asyncssh==2.4.2
bcrypt==3.2.0
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
ciscoconfparse==1.5.19
colorama==0.4.4
cryptography==3.2.1
dnspython==2.0.0
future==0.18.2
idna==2.10
Jinja2==2.11.2
junos-eznc==2.5.4
lxml==4.6.1
MarkupSafe==1.1.1
mypy-extensions==0.4.3
napalm==3.2.0
ncclient==0.6.9
netaddr==0.8.0
netmiko==3.3.2
nornir==3.0.0
nornir-jinja2==0.1.0
nornir-napalm==0.1.1
nornir-netmiko==0.1.1
nornir-scrapli==2020.11.1
nornir-utils==0.1.1
ntc-templates==1.6.0
packaging==20.4
paramiko==2.7.2
passlib==1.7.4
pkg-resources==0.0.0
pycparser==2.20
pyeapi==0.8.3
PyNaCl==1.4.0
pyparsing==2.4.7
pyserial==3.4
PyYAML==5.3.1
requests==2.24.0
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
scp==0.13.3
scrapli==2020.10.10
scrapli-asyncssh==2020.10.10
scrapli-community==2020.9.19
scrapli-netconf==2020.10.24
six==1.15.0
soupsieve==2.0.1
tenacity==6.2.0
textfsm==1.1.0
transitions==0.8.5
typing-extensions==3.7.4.3
urllib3==1.25.11
yamlordereddictloader==0.4.0
